+++ packages/stream_core_flutter/lib/src/components/badge/stream_online_indicator.dart
@@ -199,7 +199,8 @@ class DefaultStreamOnlineIndicator extends StatelessWidget {
       children: [
         props.child!,
         Positioned.fill(
-          child: Align(
+          child: FittedBox(
+            fit: BoxFit.none,
             alignment: effectiveAlignment,
             child: Transform.translate(
               offset: effectiveOffset,