+++ packages/stream_core_flutter/lib/src/components/common/stream_text_input.dart
@@ -424,7 +424,6 @@ class _DefaultStreamTextInputState extends State<DefaultStreamTextInput> {
               color: effectiveIconColor,
             ),
             child: StreamColumn(
-              mainAxisSize: .min,
               spacing: spacing.sm,
               crossAxisAlignment: .start,
               children: [