+++ packages/stream_core_flutter/lib/src/components/message/stream_message_content.dart
@@ -180,7 +180,6 @@ class DefaultStreamMessageContent extends StatelessWidget {
     return SizedBox(
       width: double.infinity,
       child: StreamColumn(
-        mainAxisSize: .min,
         spacing: effectiveSpacing,
         crossAxisAlignment: crossAxisAlignment,
         children: [?props.header, props.child, ?props.footer],